整理代码，把"我的" 全部 移动到 mine 这个类里面

+++ src/main/java/com/jfeat/am/module/cr/api/crud/IssueEndpoint.java
@@ -1,6 +1,7 @@
 package com.jfeat.am.module.cr.api.crud;
 
 import com.jfeat.am.core.jwt.JWTKit;
+import com.jfeat.am.core.shiro.ShiroKit;
 import com.jfeat.am.module.cr.services.crud.filter.IssueFilter;
 import com.jfeat.am.module.cr.services.definition.IssueStatus;
 import com.jfeat.am.module.cr.services.persistence.model.Issue;
@@ -84,6 +85,10 @@ public class IssueEndpoint{
         entity.setId(id);
         IssueFilter filter = new IssueFilter();
         filter.ignore(false);
+        Issue issue = issueService.retrieveMaster(id);
+        if (!issue.getCreateBy().equals(JWTKit.getUserId())){
+            throw new BusinessException(5000,"请勿对他人创建的issue进行修改，如有需要，请提交说明！");
+        }
         return SuccessTip.create(issueService.updateMaster(entity,filter));
     }
 
@@ -91,6 +96,9 @@ public class IssueEndpoint{
     @DeleteMapping("/{id}")
     @ApiOperation("删除 Issue")
     public Tip deleteIssue(@PathVariable Long id) {
+        if (!ShiroKit.hasRole("admin")){
+            throw new BusinessException(5100,"普通用户无法执行管理员操作!");
+        }
         return SuccessTip.create(issueService.deleteMaster(id));
     }
